+++ src/client.js
@@ -1,4 +1,5 @@
 import 'babel-polyfill'
+import 'normalize.css'
 import DeviceProvider from './components/DeviceProvider'
 import MobileDetect from 'mobile-detect'
 import React from 'react'
@@ -8,8 +9,111 @@ import configureStore from './store/configureStore'
 import createRoutes from './routes'
 import { Provider } from 'react-redux'
 import { Router } from 'react-router'
+import { colors, layout, letterSpace, lineHeight, typography } from './themes/common-variables'
+import { injectGlobal } from 'styled-components'
 import { match, browserHistory } from 'react-router'
+import { screen as mq } from './themes/screen'
+import { setupTokenInLocalStorage, deletAuthInfoAction, authUserByTokenAction, keys } from '@twreporter/registration'
 import { syncHistoryWithStore } from 'react-router-redux'
+
+// inject global styles into html
+injectGlobal`
+  html {
+    font-size: ${typography.font.size.base};
+    ::selection {
+      background-color: ${colors.red.lightRed};
+      color: $FFF;
+    }
+  }
+  body {
+    letter-spacing: ${letterSpace.generalLetterSpace};
+
+    line-height: ${lineHeight.lineHeightMedium};
+
+    font-family: "source-han-sans-traditional", "Noto Sans TC", "PingFang TC", "Apple LiGothic Medium", Roboto, "Microsoft JhengHei", "Lucida Grande", "Lucida Sans Unicode", sans-serif;
+
+    abbr[title], abbr[data-original-title] {
+      border-bottom: 0;
+    }
+
+    *, :before, :after {
+      box-sizing: border-box;
+    }
+
+    a {
+      text-decoration: none;
+    }
+
+    img {
+      vertical-align: middle;
+    }
+
+    .hidden {
+      display: none !important;
+    }
+
+    .container {
+      line-height: ${lineHeight.linHeightLarge};
+    }
+
+    .inner-max {
+      ${mq.desktopAbove`
+        max-width: ${layout.article.innerWidth};
+      `}
+    }
+
+    .outer-max {
+      ${mq.desktopAbove`
+        max-width: ${layout.article.outerWidth};
+      `}
+    }
+
+    .no-hover {
+        border-bottom: 0 !important;
+        &:after {
+            display: none;
+        }
+        &:hover:after {
+            width: 0;
+            display: none;
+        }
+    }
+
+    .text-justify {
+      text-align: justify;
+    }
+
+    .text-center {
+      text-align: center;
+    }
+
+    .center-block {
+      display:block;
+      margin-left:auto;
+      margin-right:auto;
+    }
+
+    .visible-print {
+      display: none;
+    }
+
+    figure, p {
+      margin: 0;
+    }
+
+    @media print {
+      .hidden-print {
+        display: none !important;
+      }
+      .visible-print {
+        display: block !important;
+      }
+      a[href]:after {
+        content: '';
+      }
+    }
+  }
+`
 
 let reduxState
 if (window.__REDUX_STATE__) {
@@ -28,6 +132,23 @@ if (window.__REDUX_STATE__) {
 const store = configureStore(browserHistory, reduxState)
 
 const history = syncHistoryWithStore(browserHistory, store)
+
+// token can be stored in localStorage in two scenario
+// 1. TWReporter account sign in
+// 2. oAuth
+// Acount: store auth info during signin action
+// oAuth: cookie -> redux state -> localStorage -> delete authinfo in redux state
+// The following procedure is only for oAuth
+const { auth } = store.getState()
+if(auth.authenticated && auth.authInfo && (auth.authType=== 'facebook' || auth.authType==='google')) {
+  setupTokenInLocalStorage(auth.authInfo, keys.LOCALSTORAGE_KEY_AUTH)
+  store.dispatch(deletAuthInfoAction())
+}
+
+// Check if token existed in localStorage and expired
+// following preocedure is for both accoutn and oAuth SignIn
+// 30 = 30 days
+store.dispatch(authUserByTokenAction(30, auth.authType))
 
 const device = store.getState().device